+++ web-client/src/components/forms/Button/Button.rendering.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react"
 import { Button } from "./Button.component"
 
 describe("Button.component: rendering", () => {
-  it("renders an Button component with a defined text", () => {
+  it("renders a Button component with a defined text", () => {
     const label = "Click me"
     render(<Button>{label}</Button>)
     const buttonElement = screen.getByRole("button")