+++ web-client/src/utils/helpers/types.helpers.ts
@@ -1,0 +1,5 @@
+export function assertIsNode(event: EventTarget | null): asserts event is Node {
+  if (!event || !(event instanceof Node)) {
+    throw new Error("Node expected")
+  }
+}